Changing Point x and y by integers

+++ src/model/Point.java
@@ -2,29 +2,29 @@ package model;
 
 public class Point {
 
-    private float x;
-    private float y;
+    private int x;
+    private int y;
 
-    public Point(float x, float y) {
+    public Point(int x, int y) {
         this.x = x;
         this.y = y;
     }
 
-    public float getX() {
+    public int getX() {
         return x;
     }
 
-    public void setX(float x) {
+    public void setX(int x) {
         this.x = x;
     }
 
 
-    public float getY() {
+    public int getY() {
         return y;
     }
 
 
-    public void setY(float y) {
+    public void setY(int y) {
         this.y = y;
     }
